Extending behavior of VCXProject.
Automatically fill the .OutputDirectory attribute
if not specified from the VCXProject target node.

Automatically fill the .IntermediateDirectory attribute
if not specified from the VCXProject target node.

Consider a CSNode in the same manner as an (e.g.) DLLNode or
an ExecuteNode when determining the .OutputDirectory attribute
and the .LocalDebuggerCommand attribute.